+++ test/service/configInjection.test.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import { expect } from 'chai';
 
-import { prepareInjectionConfigVars, convertToKeyValuePairs } from '../src/services/configInjection';
+import { prepareInjectionConfigVars, convertToKeyValuePairs } from '../../src/service/configInjection';
 
 describe('Services: configInjection', () => {
   describe('prepareInjectionConfigVars', () => {
